Added testing sequence to mira.c to see if instrument measures properly and returns data as expected.

+++ miracode/Core/Inc/mira.h
@@ -35,7 +35,7 @@ extern const uint8_t POWERSAVE;
 extern const uint8_t EMPTY_PAYLOAD; //Check that this is not used
 
 //MIRA communication status
-extern volatile uint8_t mira_ready_for_comm;
+extern volatile unsigned mira_ready_for_comm;
 
 HAL_StatusTypeDef mira_command(UART_HandleTypeDef *huart, uint8_t command, uint8_t reg, uint32_t *data, uint8_t *rxBuffer, uint32_t Timeout);
 HAL_StatusTypeDef mira_science_data(UART_HandleTypeDef *huart, uint8_t *science_Rx, uint8_t *response_Rx, uint32_t Timeout);
